Add changes from wedding. Works great! https://xkcd.com/1296/

+++ kivybooth.sh
@@ -15,8 +15,10 @@ cd $(dirname $0)
 python main.py 2>&1 | grep --line-buffered --invert-match '^INFO: The key you just pressed is not recognized by SDL\.'
 rc=${PIPESTATUS[0]}
 
-if [ $rc -ne 0 ] && [ $rc -ne 120 ] ; then # Check if we stopped unintentionally
+if [ $# -eq 0 ] && [ $rc -ne 0 ] && [ $rc -ne 120 ] ; then # Check if we stopped unintentionally
     echo "Kivybooth failed (RC=${rc}). Will restart system in 30 seconds..."
     sleep 30 # Wait a little so that we have a little time to connect to the device in case of boot loops
     reboot
 fi
+
+exit $rc
